Vcoi a imag

+++ Pyxi/DataAcquisition.py
@@ -92,7 +92,7 @@ class DataAcquisitionThread(Qt.QThread):
     def OutSignal(self, Amp):
         step = 2*np.pi*(self.Freq/self.FsScope)
         self.Signal = np.float64(Amp*np.exp(1j*(step*np.arange(self.EveryN))))
-        self.Vcoi = np.float64(1*np.exp(1j*(step*np.arange(self.EveryN))))
+        self.Vcoi = 1*np.exp(1j*(step*np.arange(self.EveryN)))
 
     def run(self, *args, **kwargs):
         self.DaqInterface.StartAcquisition(Fs=self.FsScope,
